requirements.txt
loguru
speedtest-cli